+++ src/main/java/wooteco/auth/web/AuthenticationPrincipalArgumentResolver.java
@@ -6,7 +6,6 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import wooteco.auth.domain.AuthenticationPrincipal;
 import wooteco.auth.infrastructure.AuthorizationExtractor;
 import wooteco.auth.infrastructure.JwtTokenProvider;
